Identifying container images by registry, name and digest
The image-ref is not a reliable source to identify images when it comes
from the docker daemon (in the docker://... form). This will identify
the images from the information that we can parse.

Doing so will also enable us to commit the images to the @data hash when
we identify the images instead from a collection function (get_images),
which will simplify image collection from Openshift.